Move components Footer, Nav to subpaths
